+++ apps/doc/src/routes/(rails)/doc/decisions/+page.svelte
@@ -23,20 +23,18 @@
 	layout="tram"
 >
 	{#snippet main()}
-		<div class="l:text:md">
-			<ul class="l:grid:auto size:sm">
-				{#each markdowns as { meta }, i (i)}
-					{#if meta.status !== 'draft'}
-						<a
-							class="surface:2:neutral ravioli:sm"
-							href={`${page.url.pathname}/${meta.slug}`}
-						>
-							{Number(meta.id)} - {meta.title}
-						</a>
-					{/if}
-				{/each}
-			</ul>
-		</div>
+		<ul class="l:grid:auto size:sm">
+			{#each markdowns as { meta }, i (i)}
+				{#if meta.status !== 'draft'}
+					<a
+						class="surface:2:neutral ravioli:sm"
+						href={`${page.url.pathname}/${meta.slug}`}
+					>
+						{Number(meta.id)} - {meta.title}
+					</a>
+				{/if}
+			{/each}
+		</ul>
 	{/snippet}
 	{#snippet aside()}
 		<p class="font:md font:heading font:semibold cannelloni:inline size:2xs">